Consolidate helpers in re_data_store::util
